add dao impl

+++ tuomatuo-core/src/main/java/com/lami/tuomatuo/dao/DynamicCommentDaoInterface.java
@@ -6,5 +6,5 @@ import com.lami.tuomatuo.model.DynamicComment;
 /**
  * Created by xujiankang on 2016/1/18.
  */
-public interface DynamicCommentDaoInterface extends BaseDao<DynamicComment, Integer> {
+public interface DynamicCommentDaoInterface extends BaseDao<DynamicComment, Long> {
 }
