+++ tuomatuo-core/src/main/java/com/lami/tuomatuo/dao/QQAccountDaoInterface.java
@@ -6,5 +6,5 @@ import com.lami.tuomatuo.model.QQAccount;
 /**
  * Created by xujiankang on 2016/1/18.
  */
-public interface QQAccountDaoInterface extends BaseDao<QQAccount, Integer> {
+public interface QQAccountDaoInterface extends BaseDao<QQAccount, Long> {
 }
